+++ main.go
@@ -17,7 +17,6 @@
 package main
 
 import (
-	"fmt"
 	"os"
 	"strings"
 
@@ -25,13 +24,18 @@ import (
 	"github.com/containerd/containerd/defaults"
 	"github.com/containerd/containerd/namespaces"
 	ncdefaults "github.com/containerd/nerdctl/pkg/defaults"
-	"github.com/containerd/nerdctl/pkg/infoutil"
 	"github.com/containerd/nerdctl/pkg/logging"
 	"github.com/containerd/nerdctl/pkg/rootlessutil"
 	"github.com/containerd/nerdctl/pkg/version"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
+)
+
+type Category = string
+
+const (
+	CategoryManagement = Category("Management")
 )
 
 func main() {
@@ -69,7 +73,7 @@ func newApp() *cli.App {
 			Aliases: []string{"a", "host", "H"},
 			Usage:   "containerd address, optionally with \"unix://\" prefix",
 			EnvVars: []string{"CONTAINERD_ADDRESS"},
-			Value:   "unix://" + defaults.DefaultAddress, // same for rootless as well (mount-namespaced)
+			Value:   defaults.DefaultAddress, // same for rootless as well (mount-namespaced)
 		},
 		&cli.StringFlag{
 			Name:    "namespace",
@@ -181,90 +185,4 @@ func newApp() *cli.App {
 	}
 	app.BashComplete = appBashComplete
 	return app
-}
-
-func appNeedsRootlessParentMain(clicontext *cli.Context) bool {
-	if !rootlessutil.IsRootlessParent() {
-		return false
-	}
-	// TODO: allow `nerdctl <SUBCOMMAND> --help` without nsentering into RootlessKit
-	switch clicontext.Args().First() {
-	case "", "completion", "login", "logout":
-		return false
-	}
-	return true
-}
-
-type Category = string
-
-const (
-	CategoryManagement = Category("Management")
-)
-
-func appBashComplete(clicontext *cli.Context) {
-	w := clicontext.App.Writer
-	coco := parseCompletionContext(clicontext)
-	switch coco.flagName {
-	case "n", "namespace":
-		bashCompleteNamespaceNames(clicontext)
-		return
-	case "snapshotter", "storage-driver":
-		bashCompleteSnapshotterNames(clicontext)
-		return
-	case "cgroup-manager":
-		fmt.Fprintln(w, "cgroupfs")
-		if ncdefaults.IsSystemdAvailable() {
-			fmt.Fprintln(w, "systemd")
-		}
-		if rootlessutil.IsRootless() {
-			fmt.Fprintln(w, "none")
-		}
-		return
-	}
-	cli.DefaultAppComplete(clicontext)
-	for _, subcomm := range clicontext.App.Commands {
-		fmt.Fprintln(clicontext.App.Writer, subcomm.Name)
-	}
-}
-
-func bashCompleteNamespaceNames(clicontext *cli.Context) {
-	if rootlessutil.IsRootlessParent() {
-		_ = rootlessutil.ParentMain()
-		return
-	}
-
-	client, ctx, cancel, err := newClient(clicontext)
-	if err != nil {
-		return
-	}
-	defer cancel()
-	nsService := client.NamespaceService()
-	nsList, err := nsService.List(ctx)
-	if err != nil {
-		logrus.Warn(err)
-		return
-	}
-	for _, ns := range nsList {
-		fmt.Fprintln(clicontext.App.Writer, ns)
-	}
-}
-
-func bashCompleteSnapshotterNames(clicontext *cli.Context) {
-	if rootlessutil.IsRootlessParent() {
-		_ = rootlessutil.ParentMain()
-		return
-	}
-
-	client, ctx, cancel, err := newClient(clicontext)
-	if err != nil {
-		return
-	}
-	defer cancel()
-	snapshotterPlugins, err := infoutil.GetSnapshotterNames(ctx, client.IntrospectionService())
-	if err != nil {
-		return
-	}
-	for _, name := range snapshotterPlugins {
-		fmt.Fprintln(clicontext.App.Writer, name)
-	}
 }
