+++ pkg/infoutil/infoutil_linux_test.go
